PROJ: Generic Coordinate System Transformations Using 'PROJ'

A wrapper around the generic coordinate transformation software 'PROJ' that transforms geospatial coordinates from one coordinate reference system ('CRS') to another. This includes cartographic projections as well as geodetic transformations. Version 6.0.0 or higher is required, earlier versions if available are not used leaving this package harmlessly non-functional. The intention is for this package to be used by user-packages such as 'reproj', and that the older 'PROJ.4' and version 5 pathways be provided by the 'proj4' package. Separating this disruptive version change (from 4.0 and 5.0, to 6.0 and above) allows the use of existing and stable code in 'proj4' alongside the new idioms and requirements of modern 'PROJ' using this package. The 'PROJ' library is available from <https://proj.org/>.

The goal of PROJ is to provide generic coordinate system transformations in R without requiring bespoke formats for storing basic data.

I. Functions

See also

Author

Maintainer: Michael D. Sumner mdsumner@gmail.com (ORCID)

Other contributors:

Jeroen Ooms (provided PROJ library support on Windows, and assistance with Windows configuration) [contributor]

Simon Urbanek (wrote original code versions for PROJ version 6) [copyright holder, contributor]