Generate a projection string.

Input any accepted format of PROJ crs specification. Return value is a string in the requested format.

proj_create(source, format = 0L)

Arguments

source
input projection specification (PROJ4, WKT2, EPSG, PROJJSON, ...)

format
integer, 0 for WKT, 1 for PROJ, 2 for PROJJSON

Value

character string in requested format

Details

See the library documentation for details on input and output formats.

Some nuances of the format are not available, currently we use formats '0: PJ_WKT2_2018', '1: PJ_PROJ_5', '2: PROJJSON'.

Some formats are hard to read, such as WKT for easy reading use cat().

Examples

if (ok_proj6()) {
proj_create("EPSG:4326", format = 1)

proj_create("urn:ogc:def:crs:EPSG::4326")

proj_create("urn:ogc:def:crs:EPSG::4326", format = 1L)

cat(wkt <- proj_create("EPSG:3857"))
proj_create(wkt, format = 1L)

wkt_method <- proj_create("+proj=etmerc +lat_0=38 +lon_0=125 +ellps=bessel")

cat(wkt_method)
proj_create(wkt_method, format = 1L)
s1 <- "+proj=merc +a=6378137 +b=6378137 +lat_ts=0 +lon_0=0 +x_0=0"
s2 <- "+y_0=0 +k=1 +units=m +nadgrids=@null +wktext +no_defs +type=crs"
cat(proj_create(paste(s1, s2)))

cat(proj_create(wkt, format = 2L))
}

#> PROJCRS["WGS 84 / Pseudo-Mercator",
#>     BASEGEOGCRS["WGS 84",
#>         DATUM["World Geodetic System 1984",
#>             ELLIPSOID["WGS 84",6378137,298.257223563,
#>                 LENGTHUNIT["metre",1]]],
#>         PRIMEM["Greenwich",0,
#>             ANGLEUNIT["degree",0.0174532925199433]],
#>         ID["EPSG",4326]],
#>     CONVERSION["Popular Visualisation Pseudo-Mercator",
#>         METHOD["Popular Visualisation Pseudo Mercator",
#>             ID["EPSG",1024]],
#>         PARAMETER["Latitude of natural origin",0,
#>             ANGLEUNIT["degree",0.0174532925199433],
#>             ID["EPSG",8801]],
#>         PARAMETER["Longitude of natural origin",0,
#>             ANGLEUNIT["degree",0.0174532925199433],
#>             ID["EPSG",8802]],
#>         PARAMETER["False easting",0,
#>             LENGTHUNIT["metre",1],
#>             ID["EPSG",8806]],
#>         PARAMETER["False northing",0,
#>             LENGTHUNIT["metre",1],
#>             ID["EPSG",8807]]],
#>     CS[Cartesian,2],
#>         AXIS["easting (X)",east,
#>             ORDER[1],
#>             LENGTHUNIT["metre",1]],
#>         AXIS["northing (Y)",north,
#>             ORDER[2],
#>             LENGTHUNIT["metre",1]],
#>     USAGE[
#>         SCOPE["unknown"],
#>         AREA["World - 85°S to 85°N"],
#>         BBOX[-85.06,-180,85.06,180]],
#>     ID["EPSG",3857]]CONVERSION["PROJ-based coordinate operation",
#>     METHOD["PROJ-based operation method: +proj=etmerc +lat_0=38 +lon_0=125 +ellps=bessel"]]PROJCRS["WGS 84 / Pseudo-Mercator",
#>     BASEGEOGCRS["WGS 84",
#>         DATUM["World Geodetic System 1984",
#>             ELLIPSOID["WGS 84",6378137,298.257223563,
#>                 LENGTHUNIT["metre",1]]],
#>         PRIMEM["Greenwich",0,
#>             ANGLEUNIT["degree",0.0174532925199433]],
#>         ID["EPSG",4326]],
#>     CONVERSION["unnamed",
#>         METHOD["Popular Visualisation Pseudo Mercator",
#>             ID["EPSG",1024]],
#>         PARAMETER["Latitude of natural origin",0,
#>             ANGLEUNIT["degree",0.0174532925199433],
#>             ID["EPSG",8801]],
#>         PARAMETER["Longitude of natural origin",0,
#>             ANGLEUNIT["degree",0.0174532925199433],
#>             ID["EPSG",8802]],
#>         PARAMETER["False easting",0,
#>             LENGTHUNIT["metre",1],
#>             ID["EPSG",8806]],
#>         PARAMETER["False northing",0,
#>             LENGTHUNIT["metre",1],
#>             ID["EPSG",8807]]],
#>     CS[Cartesian,2],
#>         AXIS["(E)",east,
#>             ORDER[1],
#>             LENGTHUNIT["metre",1,
#>                 ID["EPSG",9001]]],
#>         AXIS["(N)",north,
#>             ORDER[2],
#>             LENGTHUNIT["metre",1,
#>                 ID["EPSG",9001]]]]{
#>   "$schema": "https://proj.org/schemas/v0.1/projjson.schema.json",
#>   "type": "ProjectedCRS",
#>   "name": "WGS 84 / Pseudo-Mercator",
#>   "base_crs": {
#>     "name": "WGS 84",
#>     "datum": {
#>       "type": "GeodeticReferenceFrame",
#>       "name": "World Geodetic System 1984",
#>       "ellipsoid": {
#>         "name": "WGS 84",
#>         "semi_major_axis": 6378137,
#>         "inverse_flattening": 298.257223563
#>       }
#>     },
#>     "coordinate_system": {
#>       "subtype": "ellipsoidal",
#>       "axis": [
#>         {
#>           "name": "Latitude",
#>           "abbreviation": "lat",
#>           "direction": "north",
#>           "unit": "degree"
#>         },
#>         {
#>           "name": "Longitude",
#>           "abbreviation": "lon",
#>           "direction": "east",
#>           "unit": "degree"
#>         }
#>       ]
#>     },
#>     "id": {
#>       "authority": "EPSG",
#>       "code": 4326
#>     }
#>   },
#>   "conversion": {
#>     "name": "unnamed",
#>     "method": {
#>       "name": "Popular Visualisation Pseudo Mercator",
#>       "id": {
#>         "authority": "EPSG",
#>         "code": 1024
#>       }
#>     },
#>     "parameters": [
#>       {
#>         "name": "Latitude of natural origin",
#>         "value": 0,
#>         "unit": "degree",
#>         "id": {
#>           "authority": "EPSG",
#>           "code": 8801
#>         }
#>       },
#>       {
#>         "name": "Longitude of natural origin",
#>         "value": 0,
#>         "unit": "degree",
#>         "id": {
#>           "authority": "EPSG",
#>           "code": 8802
#>         }
#>       },
#>       {
#>         "name": "False easting",
#>         "value": 0,
#>         "unit": "metre",
#>         "id": {
#>           "authority": "EPSG",
#>           "code": 8806
#>         }
#>       },
#>       {
#>         "name": "False northing",
#>         "value": 0,
#>         "unit": "metre",
#>         "id": {
#>           "authority": "EPSG",
#>           "code": 8807
#>         }
#>       }
#>     ]
#>   },
#>   "coordinate_system": {
#>     "subtype": "Cartesian",
#>     "axis": [
#>       {
#>         "name": "Easting",
#>         "abbreviation": "E",
#>         "direction": "east",
#>         "unit": "metre"
#>       },
#>       {
#>         "name": "Northing",
#>         "abbreviation": "N",
#>         "direction": "north",
#>         "unit": "metre"
#>       }
#>     ]
#>   },
#>   "scope": "unknown",
#>   "area": "World - 85°S to 85°N",
#>   "bbox": {
#>     "south_latitude": -85.06,
#>     "west_longitude": -180,
#>     "north_latitude": 85.06,
#>     "east_longitude": 180
#>   },
#>   "id": {
#>     "authority": "EPSG",
#>     "code": 3857
#>   }
#> }